+++ frontend/app/lib/main.dart
@@ -4,7 +4,7 @@ import 'package:app/view/calender_app_bar.dart';
 import 'package:app/view/calender_menu.dart';
 import 'package:app/view/calender_view/day_calender.dart';
 import 'package:app/view/calender_view/week_month_calender.dart';
-import 'package:app/view/login.dart';
+import 'package:app/view/sign_in.dart';
 import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:intl/date_symbol_data_local.dart';
@@ -39,7 +39,7 @@ class MyApp extends StatelessWidget {
               return ChangeNotifierProvider(
                   create: (context) => DateProvider(), child: MyHomePage());
             } else if (snapshot.data == false) {
-              return Login();
+              return SignIn();
             } else {
               return Scaffold(
                 body: Text(
